requirements.txt
Flask
flask_sqlalchemy
pdfplumber
pythainlp
sentence-transformers
faiss-cpu
requests
requests-oauthlib